requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
logzero==1.7.0
lxml==4.9.3
outcome==1.2.0
packaging==23.1
pycodestyle==2.10.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
trio==0.22.2
trio-websocket==0.10.3
tzdata==2022.7
urllib3==2.0.4
webdriver-manager==4.0.0
wsproto==1.2.0
